Backend/requirements.txt
pip
flask~=3.1.1
spotipy~=2.25.1
requests~=2.32.3
pandas
python-dotenv~=1.1.0
pymongo~=4.13.0
flask-cors~=6.0.0
tenacity~=9.1.2
uvicorn~=0.34.2
fastapi~=0.115.12
pydantic~=2.11.5